feat: board컴포넌트 suspense 적용
