+++ client/src/components/board/Loading/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Loading() {
+  return (
+    <div className="flex flex-col max-w-4xl m-auto gap-14 mb-4 bg-red-300">
+      Loading....
+    </div>
+  );
+}
+
+export default Loading;
