+++ client/src/components/board/index.tsx
@@ -1,0 +1,2 @@
+export { default as PostContent } from './PostContent';
+export { default as Loading } from './Loading';
